requirements.txt
numpy==1.21.2
Wave==0.0.2
numpy==1.21.2
progressbar2==3.53.3
python-utils==2.5.6
six==1.16.0
Wave==0.0.2
